Moved spatial partitioning logic to new file.
